Update download_mod to new conventions
1. Now utilizes `common.check_local_mod_paths`
2. `destination` is no longer the direct path to the mod; adding
`plugin_dest`.

+++ src/modules/providers/ThunderstoreProvider.py
@@ -3,29 +3,36 @@ import os
 import zipfile
 import requests
 from modules.providers.ProviderAbstract import ProviderAbstract
+from modules.updater.common import check_local_mod_paths
 
 
 class ThunderstoreProviderBase(ProviderAbstract):
-    def download_mod(self, log, mod_data, destination):
-        plugin = mod_data['name']
+    def download_mod(self, log, mod_data, local_paths, destination):
+        plugin = mod_data
         plugin_url = f"https://thunderstore.io/package/download/{plugin.replace('-', '/')}"
-        plugin_zip = destination + '.zip'
+        plugin_dest = os.path.join(destination, plugin)
+        plugin_zip = plugin_dest + '.zip'
 
         req = requests.get(plugin_url, allow_redirects=True)
+
+        if check_local_mod_paths(log, local_paths, destination, plugin):
+            return
 
         if req.status_code == 200:
             log(f'[INFO] (D) {plugin}')
             open(plugin_zip, 'wb').write(req.content)
 
             with zipfile.ZipFile(plugin_zip, 'r') as ref:
-                ref.extractall(destination)
+                ref.extractall(plugin_dest)
 
             os.remove(plugin_zip)
         else:
             raise Exception(f'Invalid response from Thunderstore while downloading plugin: {plugin}.')
         
+
     def move_custom_mods(self, mods_dir='', variables={}, mod_index=[], ignore=[]):
         raise NotImplementedError
+
 
     # The second parameter of the abstract method, game, is unused, as denoted by _.
     def get_latest_modpack_version(self, _, modpack):
@@ -43,18 +50,22 @@ class ThunderstoreProviderBase(ProviderAbstract):
             'url': content['latest']['download_url']
         }
     
+
     def download_modpack(self, variables):
         req = super().download_modpack(variables)
 
         if req.status_code != 200:
             raise Exception(f'Invalid from Thunderstore while downloading modpack: {variables['version']['name']}.')
-        
+
+
     def extract_modpack(self, variables, game, pack):
         return super().extract_modpack(variables, game, pack)
     
+
     def get_modpack_modlist(self, variables):
         raise NotImplementedError
     
+
     def initial_modpack_install(self, variables):
         raise NotImplementedError
 
@@ -68,11 +79,13 @@ class ThunderstoreValheimProvider(ThunderstoreProviderBase):
 
         return super().move_custom_mods(plugins_dir, variables, mod_index, ignore)
 
+
     def get_modpack_modlist(self, variables):
         manifest_json_path = os.path.join(variables.get('tmp'), 'manifest.json')
         contents = open(manifest_json_path, 'r').read()
         dependencies = json.loads(contents).get('dependencies')
         return dependencies
+
 
     def initial_modpack_install(self, variables):
         pass
